upgrade rails to 6.1 (#15)
@spacechurro I looked at this and everything is good; I generated a new app and that was all good too, so I'm merging.

I added this: https://github.com/Beyond-Finance/rails-template-generator/pull/16

Just to enable the autostart y/n query on generated apps, looks like the development environment boilerplate file grew past the rubocop default line length limit. Everything is A-OK as it is, that just adds the nicety for the script of launching the new app (instead of rubocop exiting the process right before that).